Rename tests/gdbmtool to tests/dejagnu.
Reportedly, Cygwin is unable to correctly handle the case when a directory
and a binary of the same name reside in two different PATH components,
and the component where the directory resides is listed first.  This causes
hard to diagnose errors when running the testsuite, because tests/gdbmtool
(the directory), appears before tools/gdbmtool (the binary).

The case was reported by Marco Atzeri <marco.atzeri@gmail.com>.